ai-service/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    """Configuration for AI Question Generation Engine"""
    
    # API Configuration
    API_HOST: str = os.getenv("API_HOST", "0.0.0.0")
    API_PORT: int = int(os.getenv("API_PORT", "8000"))
    DEBUG: bool = os.getenv("DEBUG", "False").lower() == "true"
    
    # AI Configuration - GEMINI AU LIEU D'OPENAI
    GEMINI_API_KEY: str = os.getenv("GEMINI_API_KEY", "")
    AI_MODEL: str = os.getenv("AI_MODEL", "gemini-pro")
    
    # External Services
    NODE_BACKEND_URL: str = os.getenv("NODE_BACKEND_URL", "http://localhost:5000")

settings = Settings()